Add demo of GetChildWorkflowExecution issue
When a child workflow is mocked using OnWorkflow,
GetChildWorkflowExecution's does not work as expected. The returned
future, when Get is called appears to block until the mocked
child workflow is completely resolved to a return value, ending the
completing the child workflow's execution. This completion of the
mocked child workflow includes the testsuite skipping time forward if
the mocked child workflow set any wait duration with After or AfterFn.

A test suite is provided with tests which are all expected to PASS, but
two tests FAIL due to the described unexpected behavior.